api_bancaria/pyproject.toml
[project]
name = "api-bancaria"
version = "0.1.0"
description = ""
authors = [
    {name = "V1n1c1u5P4iva",email = "paivavinicius214@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.39,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
